Added more code for use as DLL
This commit adds more code for using LIL as a DLL. Programs that wish to use LIL as DLL must compile LIL with the LILDLL definition. Their code must also be compiled with the LILDLL definition and their callbacks with LILCALLBACK modifier. Existing programs or programs which do not wish to use LIL as a DLL can ignore these.

The commit also adds OpenWatcom target and project files for making a lil.dll file.