Added custom data in lil_t objects
This commit adds lil_set_data and lil_get_data which can be used to associate custom data (via a single pointer) to a lil_t object. This is useful in the case where there are multiple lil_t objects lying around which are encapsulated in some outer object (like a C++ object) and a callback function needs to figure out the outer object from the lil_t object.