Added the ability to "catch" unknown function calls in scripts
This commit adds a new function, catcher, which can be used to set and retrieve LIL code which will be called when a function call fails because the function name was not known. The code is executed in place of the function (in a similar way to anonymous functions) inside its own environment. Catchers can be used to implement embedded DSLs or shell (as in UNIX shell) like functionality. The commit also adds an example program (catcher.lil) and updates the readme.txt file with information about the new functionality.

Also the commit updates readme.txt to use double spaces everywhere.