Added reflect dollar-prefix
This commit adds a new subcommand in reflect: reflect dollar-prefix. This can be used to obtain or set the dollar prefix. Previously LIL had the 'set ' prefix hardcoded when the dollar $ was used (for variable expansion like $foo, $bar, etc). Now the dollar prefix can be altered with this new command.