Added filter and slice functions and modified eval's behavior
This commit adds two new functions: filter and slice. The "filter" function can be used to filter a list against some expression and returns the filtered list. The "slice" function can be used to get part of a list.

The commit also modifies the way "eval" behaves: the "funclevel" argument for the call to lil_parse_value is now 0, which means that the last command's result is used instead any call to "return". Also calling "return" is now the same as if it was called by the code that called eval.