Fixed a bug in expr with space skipping
This commit fixes a bug in expression parsing (expr func and others). The bug was caused with whitespace before a second operator of the same level as the last parsed (example: x * y / z -- the "/ z" part was ignored because the space before it wasn't skipped). Now a call to ee_skipspaces is made at the end of each while loop which checks for a known operator at the same level.