Added a "writechar" function in main.c
Added a new function, writechar, in main.c. The function can be used to write a character based on its code (like writing "write [char <code>]") but is not filtered via the rest of the LIL runtime and exported directly to the stdout. This can be used main to write the zero byte at output.

This has been made part of the main.c instead of lil.c because the ability to write non-ascii, non-control characters depends on the host program (most command-line based interfaces can) and is not globally available. If a host program requires this functionality, it is trivial to add it (or just copy main.c's implementation).