Fixed a bug in expression parser for signs symbols followed by non-numbers without whitespace
This commit fixes a bug in the expression parser that made sign symbols followed by non-numbers without whitespace to produce wrong results. As an example of the issue, the expression -(3*3) produced wrong results while - ( 3 * 3 ) produced correct results. Now this is fixed. It is still recommended to use whitespaces between operators and operands though.