Added local and topeval functions
This commit adds two new functions: local and topeval

The local function can be used to make local variables (if they are not
already defined in the current environment).  This is useful for
reusable functions to make sure they will not modify global variables
with the same name.

The topeval function can be used to evaluate code at the toplevel
(global) environment.