Fixed setvar target environment and added a new function
This commit fixes the target environment for lil_set_var (it used to be
always the current environment for non LIL_SETVAR_LOCAL_NEW) and adds a
new function called enveval.  The enveval function can be used to
evaluate code inside a new environment with optional variable
copy-to/from the current one.

Additionally the result of 'return' is now set to the passed value
instead of NULL.