Added a new function, result, to specify function results
This commit adds a new function called "result" which can be used to specify the result (return value) of a function without returning from the function.  The new function can also be used to return the current result value (if set previously by calling result), which basically gives some sort of indirect access to a variable holding the function's return value.