Dependency on stdint.h is now optional so that the code compiles under BCC 5.5 and (older) MSVC
This commit makes the dependency on stdint.h to be optional so that the code will compile under Borland's free C/C++ Compiler version 5.5.1.  It replaces the use of int64_t with lilint_t which, by default, is typedef'd to int64_t unless the macros LILINT_LONGLONG or LILINT_INT64 are used, which typedef lilint_t to long long int or __int64.

Also a Makefile (Makefile.bcc) for BCC32/Borland Make is included that builds the library (liblil.lib, lil.obj) and test shell (lil.exe).

The readme.txt has been updated to reference to the new API and compiler support.